Fix event.target being wrong on drag events
Drag events will not always fire with a target of the outer draggable
element. The target can sometimes be a nested child element. This throws
off measurements and can cause the wrong element to animate.

The typical fix here would be to use event.currentTarget, which should
point towards the element that has the actual event listener, but it is
null with my setup. Probably something to do with how Preact registers
event listeners.

The fix I've worked out is just to walk up the DOM and look for the
nearest draggable element. That should work regardless of what
event.target points to.