Animate released element flying to its position
This turned out to be a massive pain. The default browser behavior is
to animate a return to position when the drag process fails but to not
animate if there is a successful drop. I might be able to take advantage
of the the built in cancel animation, but dragend does not fire until a
moment or two after the animation finishes. That means there is a flash
as the dragImage returns to position, disappears, and then the actual
element reappears.

I spent a long time messing with various bad options, and finally landed
on turning the entire document body into a valid drop zone. This allows
me to customize how the drop will animate no matter where on the page
the user drops the dragged element. I will still get the flash if the
user drops the element outside of the page entirely, but I'd have to
do something ridiculous like wait for drag events to stop coming in and
guess that a dragend is *about* to fire.