Add live funhouse mirror for zacis.me project
This was perhaps a silly idea, but I thought it would be fun if zacis.me
itself was included in the project gallery and showed a live updating
version of itself instead of a static image. I figured I could do it
with a canvas and ended up relying on an aging html2canvas library.
There were quite a few wrinkles to work out, but probably still better
than writing my own.