Add mirror to zacis.me project overlay
Took a bit of finagling to get this looking good on all screen sizes:

- The overlay-background needs to stretch well above the top and bottom
  of the screen since on certain sizes, the overlay mirror will peak
  above and below
- This breaks height: auto, since it is now much too tall. Fixed with
  a wrapper that crops the image
- Need many more redraws on certain screen sizes since you can see much
  deeper into the mirror

Also, it looks like refs to Preact components are not DOM nodes. Silly.
The DOM node appears to be under the "base" property, but I had to do
some refactoring to make the overlay animations work either way.