Fix height "bobble" on overlay animation
Looks like height: auto animates poorly, with the height not changing
for the first quarter of the animation, then suddenly jumping to the
correct height. This causes a visible white splash that makes it look
like the image is out of sync with the rest of the overlay.

I need height: auto to keep the canvas from getting squashed on certain
screen sizes, but with the images I can just manually set the height and
not worry about it.