Restyle control panels and text-buttons
The grey boxes looked nice enough, but clashed with the rest of the
page. The goal of the styling is to resemble a simple resume, and the
grey boxes stood out.

Removing them also provided an opportunity to bring the anchor links and
text buttons into alignment in terms of color.