Cleanup flicker workarounds
Preact doesn't always play well with animations. One workaround for
objects that fade in is to manually set them to opacity: 0 before the
first render, then remove the style immediately afterwards when you kick
off the animation. I did this previously by setting the opacity to 1,
but that creates a lot of unnecessary style pollution. We can just
remove the opacity: 0 instead.

Also, the expanded Project should use a proper "hidden" class instead of
manually setting the opacity.