Correct AniMono tools value
When I was originally planning out the "tools" property, I thought I
would have issues including tools I didn't want to feature in the Filter
controls. Went a different way with the implementation, so I can correct
the record on AniMono (we used a bit of JavaScript, but not much).