Preload overlay images
Seeing a delay on the deployed page before the image is displayed.
Let's preload them on page load. Doing it in JS rather than in a preload
link in the HTML header, avoids having multiple sources of truth for the
srcset, and also Safari yelling at me in the console for not immediately
rendering the preloaded image.