Restyle multiline props with children
I've always been a fan of using a function like `h` (or Mithril's `m`)
rather than JSX (which to me is too verbose and different enough from
HTML to nullify any benefit of the similarity). That said, the styling
of complex `h` components can sometimes get wonky. In particular, what
to do when the props are multiple lines long and you also have children?

I have now settled on this:

```
h('tag', {
  prop1: 'prop',
  prop2: 'prop'
},
  child
)
```

This is similar to how you might style such a component in JSX:

```
<tag
  prop1="prop"
  prop2="prop"
>
  child
</tag>
```