Lock down Animated and support removals/additions
I wanted to take another pass at Animated. I want to be a little more
explicit about the API and properly support additions and removals.

The new Animated must be passed an array of children and those children
must each have a unique key. This allows me to remove a lot of fuzzy
guesswork around which child changed and how.

Moreover, all animations will now start after the next children are
rendered. This allows me to clearly compare the old locations/elements
with the newly rendered ones and stop fuzzily guessing where things are
going to end up.

The downside of this approach is that I have to get super hacky to
ensure the preact render does not briefly cause a flash before the
animations start. I accomplish this with a number of temporary style
modifications, which I undo right before kicking off the animation.

All of this allows me to properly animate the addition and removal of
children, something that was enough of a pain that I dropped the
functionality before. For my own purposes I am still going to use the
"display" prop approach for now because I like the way it flies in/out,
but I am glad have add/remove in place for the future.