Another flash of white fix
My attempts to get the overlay images to caches ahead of time do not
seem to be reliably effective. Instead I pull up the overlay hidden and
then wait for the image to load. After it has loaded, I show and animate
the overlay as before.

Still a bit of a convoluted hack, but marginally less convoluted and
MUCH more reliable.