Use dragImage instead of dragging actual element
I really liked the idea of moving the draggable element itself on drag,
it simplifies some things conceptually, but it makes other logic more
complex, especially with regards to a rendering framework like Preact.
So now I am going with an approach similar to the official Drag and Drop
API. When you drag an element a "drag image" is created which moves with
the pointer. The actual element does not move.