Allow customization for dragged element and image
My Draggable API now accepts a class which will be applied to the drag
target and a function to custom generate the drag image. The default
matches the Drag and Drop API, with the dragged element remaining
visible while the drag image is a translucent copy of the dragged
element.

Importantly, this API change will make it possible for me to fix the
issue with cloneNode not being sufficient for the Mirror element.