Final (?) flash of white fix
This has been a maddening issue to wrangle. Whether or not there is a
flash of white when you try to expand a project depends heavily on
whether or not the overlay image has been cached, and presumably every
time I have thought I have fixed it, I am in fact being tricked by
browser caching.

This time I think I have finally licked it. Probably.

1. Load in gallery projects with initial appropriately sized images
2. Once all page assets have finished loading, switch to the
   unnecessarily-large-for-the-gallery view overlay images
3. On expanding, initially render *just* the image (which should be
   cached and displayed in the gallery already), with the rest of the
   overlay transparent
4. After two frames (32ms), show the rest of the modal, hide the
   original gallery item, and animate the expansion

This seems to work. Famous last words.